+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Upcoming
+* Slight buff to unbreaking (items last ~~35%/105%/175%/245%~~ -> 50%/120%/190%/260% more)
+
 ## Beta 3.9.1
 * Farmer's delight integration
   * Changed all knives recipes to use swords instead